Rebuild release notes.
There are extra changes in the feeds, which the updated script is meant to
avoid. I think this is because I got the date format wrong in the last couple
of releases and different platforms are ordering them differently. So I don't
think it's a new bug. I'll try to fix it though. Probably by making the date
parser a bit more liberal.


[SVN r64041]