Don't use lower case category
The category view was broken because it was converted into lowercase, so
only lowercase paramater values when appropriate. Unfortunately the code
for generating the links at the top is case sensitive, so they're
slightly wrong if a parameter value has the wrong case, but that's
unlikely, and not that bad. Will need to improve this though.